requirements.txt
Flask==2.3.2
Flask-WTF==1.1.1
psycopg2-binary==2.9.6
Werkzeug==2.3.4
MarkupSafe==2.1.3
